PLay sound when score goes up
There's an issue with the game over sound that plays only when the bird is colliding with a pipe or goes over out of the top of the screen. Sound plays every time a pipe is hit. Sound doesn't play when the bird goes out from the bottom of the screen.